Improve test setup organization
