+++ tests/conftest.py
@@ -2,8 +2,14 @@
 
 from collections.abc import Generator
 from unittest.mock import AsyncMock, patch
+from pytest_homeassistant_custom_component.common import MockConfigEntry
+from custom_components.storj.const import DOMAIN
 
 import pytest
+
+
+TEST_ACCESS_GRANT = "123xyz"
+CONFIG_ENTRY_TITLE = "Storj entry title"
 
 
 @pytest.fixture
@@ -13,6 +19,17 @@ def mock_setup_entry() -> Generator[AsyncMock]:
         "custom_components.storj.async_setup_entry", return_value=True
     ) as mock_setup_entry:
         yield mock_setup_entry
+
+
+@pytest.fixture(name="mock_config_entry")
+def mock_config_entry() -> MockConfigEntry:
+    """Fixture for MockConfigEntry."""
+    return MockConfigEntry(
+        domain=DOMAIN,
+        unique_id=TEST_ACCESS_GRANT,
+        title=CONFIG_ENTRY_TITLE,
+        data={"access_grant": TEST_ACCESS_GRANT, "bucket_name": "ha-backups"},
+    )
 
 
 @pytest.fixture(autouse=True)
